+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.1
 
+* Added `be_document` matcher.
 * Added `with_default_value` method to `HaveFieldMatcher`.
 * Added `of_type` method to `HaveFieldMatcher`.
 * Added `must` and `wont` methods to `MiniTest::Spec`.